basic styling for main page done
