+++ client/src/Components/Element/Element.css
@@ -15,13 +15,14 @@
 
 .element {
     
-    max-width: 100%;
+    width: 100%;
     height: 130px;
     margin: 3rem;
  }
 
- .elementButton {
-     width: 100%;
+ .elementButtonContainer {
+     width:100%;
+     height: 100px;
      background-color: rgb(218, 179, 185);
  }
