+++ client/src/Components/Header/Navbar/NavBar.css
@@ -5,3 +5,7 @@
    
     
 }
+
+.navigationMenu {
+    background-color: rgba(201, 196, 196, 0.925);
+}
