+++ client/src/Components/Tasks/Tasks.css
@@ -1,15 +1,15 @@
 .tasksContainer {
     width: 100%;
     height: 800px;
-    background-color: black;
+    /* background-color: black; */
     overflow: scroll;
 }
 
 .taskCard {
-    max-width: 100%;
+    max-width: 95%;
     height: 180px;
-    margin: 1rem;
-    background-color: green;
+    margin: 2rem;
+    background-color: rgba(161, 159, 159, 0.925);
     display: flex;
     flex-flow: row wrap;
     justify-content: center;
@@ -17,15 +17,19 @@
 }
 
 .card_title {
-    width: 50%;
-    height: 40px;
+    width: 30%;
+    height: 100%;
     text-align: center;
     /* background-color: violet;    */
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: center;
+    align-items: center;
 }
 .leftTaskDetails {
     /* background-color: aqua; */
-    width: 50%;
-    height: 120px;
+    width: 30%;
+    height: 100%;
     margin-top: 0px;
 
 }
@@ -49,8 +53,8 @@
  .expense-date {
     display: flex;
     flex-direction: column;
-    width: 4.5rem;
-    height: 4.5rem;
+    width: 6.5rem;
+    height: 6.5rem;
     border: 1px solid #ececec;
     background-color: #2a2a2a;
     color: white;
@@ -85,24 +89,25 @@
   .rightTaskDetails {
       display: flex;
       flex-flow: row wrap;
-      width: 50%;
+      width: 35%;
+      height: 100%;
       
   }
 
-  .leftRightButton {
-      width: 50%;
-      height: 100%;
+  .deleteTaskButtonSection {
+      width: 100%;
+      height: 50px;
       display: flex;
       flex-flow: row wrap;
       justify-content: center;
       align-items: center;
       
-  }
+  } 
 
 
-  .rightRightButton {
-      width: 50%;
-      height: 100px;
+  .editTaskButtonSection {
+      width: 100%;
+      height: 50px;
       display: flex;
       flex-flow: row wrap;
       justify-content: center;
@@ -113,7 +118,7 @@
 
   .bottomTaskDetails {
       width: 100%;
-      height: 20px;
+      height: 40px;
       display: flex;
       flex-direction: row wrap;
       align-items: center;
@@ -129,3 +134,25 @@
       
 
   }
+
+  /* width */
+::-webkit-scrollbar {
+  width: 20px;
+}
+
+/* Track */
+::-webkit-scrollbar-track {
+  box-shadow: inset 0 0 5px rgba(128, 128, 128, 0.74); 
+  border-radius: 10px;
+}
+ 
+/* Handle */
+::-webkit-scrollbar-thumb {
+  background: rgba(233, 203, 203, 0.616); 
+  border-radius: 10px;
+}
+
+/* Handle on hover */
+::-webkit-scrollbar-thumb:hover {
+  background: rgba(233, 203, 203, 0.616); 
+}
