+++ client/src/pages/plan.css
@@ -1,5 +1,5 @@
 .mainPage{
-    background-color: red;
+    /* background-color: red; */
     height: 100%;
     width: 100%;
     border-radius: 10px;
